refactor: 회원가입 화면 구조를 액티비티 + 프래그먼트 방식으로 리팩토링
