+++ app/src/main/java/com/bravepeople/onggiyonggi/presentation/login/LoginActivity.kt
@@ -1,22 +1,19 @@
 package com.bravepeople.onggiyonggi.presentation.login
-import com.bravepeople.onggiyonggi.presentation.MainActivity
+
 import android.content.Intent
 import android.graphics.Color
 import android.os.Bundle
 import android.text.InputType
 import android.text.SpannableString
 import android.text.Spanned
-import android.text.TextPaint
 import android.text.method.LinkMovementMethod
 import android.text.style.ClickableSpan
 import android.view.View
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.bravepeople.onggiyonggi.R
 import com.bravepeople.onggiyonggi.databinding.ActivityLoginBinding
-import androidx.activity.addCallback
-
-
+import com.bravepeople.onggiyonggi.presentation.MainActivity
+import com.bravepeople.onggiyonggi.presentation.signup.SignUpActivity
 
 class LoginActivity : AppCompatActivity() {
     private lateinit var binding: ActivityLoginBinding
@@ -29,28 +26,6 @@ class LoginActivity : AppCompatActivity() {
 
         setupListeners()
         setupSignUpText()
-
-        onBackPressedDispatcher.addCallback(this) {
-            val fm = supportFragmentManager
-            if (fm.backStackEntryCount > 0) {
-                fm.popBackStack()
-                binding.login.visibility = View.VISIBLE
-                binding.container.visibility = View.GONE
-            } else {
-                finish()
-            }
-        }
-
-
-        binding.tvGoToSignUp.setOnClickListener {
-            binding.login.visibility = View.GONE
-            binding.container.visibility = View.VISIBLE
-
-            supportFragmentManager.beginTransaction()
-                .replace(R.id.container, SignUpAgreementFragment())
-                .addToBackStack(null)
-                .commit()
-        }
     }
 
     private fun setupListeners() {
@@ -69,9 +44,9 @@ class LoginActivity : AppCompatActivity() {
             val id = binding.etId.text.toString()
             val pw = binding.etPasswd.text.toString()
 
-            if(id == "test" && pw == "1234") {
+            if (id == "test" && pw == "1234") {
                 binding.tvLoginError.visibility = View.GONE
-                Toast.makeText(this,"로그인 성공!", Toast.LENGTH_SHORT).show()
+                Toast.makeText(this, "로그인 성공!", Toast.LENGTH_SHORT).show()
                 startActivity(Intent(this, MainActivity::class.java))
                 finish()
             } else {
@@ -82,6 +57,10 @@ class LoginActivity : AppCompatActivity() {
         binding.btnBack.setOnClickListener {
             finish()
         }
+
+        binding.tvGoToSignUp.setOnClickListener {
+            startActivity(Intent(this, SignUpActivity::class.java))
+        }
     }
 
     private fun setupSignUpText() {
@@ -90,13 +69,7 @@ class LoginActivity : AppCompatActivity() {
 
         val clickableSpan = object : ClickableSpan() {
             override fun onClick(widget: View) {
-                supportFragmentManager.beginTransaction()
-                    .replace(R.id.container, SignUpAgreementFragment())
-                    .addToBackStack(null)
-                    .commit()
-
-                findViewById<View>(R.id.login).visibility = View.GONE
-                findViewById<View>(R.id.container).visibility = View.VISIBLE
+                startActivity(Intent(this@LoginActivity, SignUpActivity::class.java))
             }
 
             override fun updateDrawState(ds: android.text.TextPaint) {
@@ -114,5 +87,4 @@ class LoginActivity : AppCompatActivity() {
         binding.tvGoToSignUp.movementMethod = LinkMovementMethod.getInstance()
         binding.tvGoToSignUp.highlightColor = android.graphics.Color.TRANSPARENT
     }
-
 }
