+++ app/src/main/java/com/bravepeople/onggiyonggi/presentation/signup/SignUpAgreementFragment.kt
@@ -1,4 +1,4 @@
-package com.bravepeople.onggiyonggi.presentation.login
+package com.bravepeople.onggiyonggi.presentation.signup
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,7 +6,6 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import com.bravepeople.onggiyonggi.R
 import com.bravepeople.onggiyonggi.databinding.FragmentSignupAgreementBinding
