+++ app/src/main/res/layout/activity_signup.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!-- res/layout/activity_signup.xml -->
+
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/signupFragmentContainer"
+    android:id="@+id/signup_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
